chore: switch back to npm from yarn 2
Yarn 2 has quite a few things going for it, for example the zero-install
mode is a nice idea. However, in a project potentially open for lots of
contributers, the benefits don't outweight the drawbacks. CI with pull
requests is harder, some tooling like dependabot cannot handle yarn 2
and the repo size blows up with the zero install model